refactor: App의 static 메서드를 util로 이동
